start_mon: use --monmap instead of --inject-monmap in mkfs
When we are going to mkfs in creating mon data, we want to inject our monmap generated local
to mon db store.

But the --inject-monmap is wrong parameter for this. If we specify mkfs, the --inject-monmap
will be ignored. Instead, in this case, we need to use --monmap.

Reference: http://docs.ceph.com/docs/mimic/rados/operations/add-or-rm-mons/

Signed-off-by: Dongsheng Yang <dongsheng.yang@easystack.cn>